Don't overwrite flipped textureTransform with unflipped on next blit
The logic introduced in 60d9509cb00526e8530926b19b2366e584fdf30a didn't
account for the fact that repeated blits with OriginTopLeft would only
hit the code path that modified the source transform to flip it if the
uniform state wasn't already IdentityFlipped.

As a result, we would end up setting the unflipped texture transform
on the next blit, even though the origin was still OriginTopLeft.

Fixes: QTBUG-98803
Change-Id: Ib19e80e026acaa43981077b98ff942a7fa060378
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>