macOS: Fix less common writing systems on Catalina and later
Since Catalina, some fonts on macOS have been put into a special
"secret" mode where they are available, but are not enumerated:
https://support.apple.com/en-in/HT210192

It appears that this is done in order to phase the fonts out, as
the article refers to backwards-compatibility as reasoning for
keeping them around, but at the same time, there does not seem to
be any new alternatives. When using e.g. TextEdit, the text is
resolved to the "secret" font, but the font dialog does not
display it, indicating that you are using a non-existent font.

This also causes issues for Qt, since we have our own fontdatabase
and do our own matching. In order to work around it, we assume the
font is available and manually add it to the database and fallback
lists. This also appears to be what others are doing, e.g. Firefox.

[ChangeLog][macOS][Text] Fixed missing text with certain writing
systems on macOS Catalina and later.

Fixes: QTBUG-96384
Fixes: QTBUG-98920
Change-Id: Ifd08ccbfdd0c06e11a44be861ad2c275c5a8e339
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit 65279d6e9dfbccf1b97884a1c19111d6ebb3828f)
Reviewed-by: Ievgenii Meshcheriakov <ievgenii.meshcheriakov@qt.io>