Revert optimizations and fixes for moving/scrolling overlapped widgets
This reverts the QtWidgets changes made in commits
22634e00794e72d68e7578e1962f9f2023870749 and
5b09346cf4322704a866f253b911d467c40df3ba while keeping the auto tests
introduced in the former commit.

Both commits introduced rendering errors when moving widgets out of
or into areas in which they are obscured. Before we apply any further
optimizations to this code we need thorough auto test coverage.

Task-number: QTBUG-98151
Task-number: QTBUG-26269
Pick-to: 6.2
Change-Id: I9cb82b73776daed59ea0e9f51ff7ddef1c7265b6
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>