QLayoutPolicy: de-inline the Policy constants
... in an attempt to fix QtQuick linker errors:

   C:/MINGW1120/mingw64/bin/../lib/gcc/x86_64-w64-mingw32/11.2.0/../../../../x86_64-w64-mingw32/bin/ld.exe: /Users/qt/work/qt/qtdeclarative/src/quicklayouts/qquicklayout.cpp:1230: undefined reference to `__imp__ZN13QLayoutPolicy5FixedE'

It seems that on Windows debug builds, inline variables, like inline
functions, are supposed to be DLL-exported, when the whole class is
exported wholesale, but actually aren't, at least on MinGW.

Fix by de-inlining the variables. It appears that we can still keep
the value in the declaration, for optimizers to constant-fold. In Qt
6.3, QLayoutPolicy isn't exported wholesale anymore, but that requires
Q_GADGET_EXPORT, which we don't have in 6.2.

Amends 3dc20d809452aa67a443ee00dc37287fd529c9b4.

Change-Id: I975417f9c0f7845d0785b4ba5e51b821c069ffc5
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>